Tim 2/Project_Akhir/monde/requirements.txt
Flask
keras==2.6.0
nltk
numpy
wordnet
tensorflow==2.6.0
scikit-learn
matplotlib==3.1.3
pillow==7.2.0
opencv-python>=4.1.2